+++ FakeNFT/Foundation/NetworkClient/NetworkClient.swift
@@ -12,7 +12,7 @@ protocol NetworkClient {
     func send(request: NetworkRequest,
               completionQueue: DispatchQueue,
               onResponse: @escaping (Result<Data, Error>) -> Void) -> NetworkTask?
-
+    
     @discardableResult
     func send<T: Decodable>(request: NetworkRequest,
                             type: T.Type,
@@ -21,13 +21,13 @@ protocol NetworkClient {
 }
 
 extension NetworkClient {
-
+    
     @discardableResult
     func send(request: NetworkRequest,
               onResponse: @escaping (Result<Data, Error>) -> Void) -> NetworkTask? {
         send(request: request, completionQueue: .main, onResponse: onResponse)
     }
-
+    
     @discardableResult
     func send<T: Decodable>(request: NetworkRequest,
                             type: T.Type,
@@ -40,7 +40,7 @@ struct DefaultNetworkClient: NetworkClient {
     private let session: URLSession
     private let decoder: JSONDecoder
     private let encoder: JSONEncoder
-
+    
     init(session: URLSession = URLSession.shared,
          decoder: JSONDecoder = JSONDecoder(),
          encoder: JSONEncoder = JSONEncoder()) {
@@ -48,7 +48,7 @@ struct DefaultNetworkClient: NetworkClient {
         self.decoder = decoder
         self.encoder = encoder
     }
-
+    
     @discardableResult
     func send(
         request: NetworkRequest,
@@ -61,18 +61,18 @@ struct DefaultNetworkClient: NetworkClient {
             }
         }
         guard let urlRequest = create(request: request) else { return nil }
-
+        
         let task = session.dataTask(with: urlRequest) { data, response, error in
             guard let response = response as? HTTPURLResponse else {
                 onResponse(.failure(NetworkClientError.urlSessionError))
                 return
             }
-
+            
             guard 200 ..< 300 ~= response.statusCode else {
                 onResponse(.failure(NetworkClientError.httpStatusCode(response.statusCode)))
                 return
             }
-
+            
             if let data = data {
                 onResponse(.success(data))
                 return
@@ -84,12 +84,12 @@ struct DefaultNetworkClient: NetworkClient {
                 return
             }
         }
-
+        
         task.resume()
-
+        
         return DefaultNetworkTask(dataTask: task)
     }
-
+    
     @discardableResult
     func send<T: Decodable>(
         request: NetworkRequest,
@@ -106,18 +106,18 @@ struct DefaultNetworkClient: NetworkClient {
             }
         }
     }
-
+    
     // MARK: - Private
-
+    
     private func create(request: NetworkRequest) -> URLRequest? {
         guard let endpoint = request.endpoint else {
             assertionFailure("Empty endpoint")
             return nil
         }
-
+        
         var urlRequest = URLRequest(url: endpoint)
         urlRequest.httpMethod = request.httpMethod.rawValue
-
+        
         if let headers = request.headers {
             for(key, value) in headers {
                 urlRequest.setValue(value, forHTTPHeaderField: key)
@@ -126,17 +126,17 @@ struct DefaultNetworkClient: NetworkClient {
         
         
         if let formData = request.dto as? [String: String] {
-                    // Transform data in x-www-form-urlencoded
-                    let formDataString = formData.map { key, value in
-                        return "\(key)=\(value.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? "")"
-                    }.joined(separator: "&")
-                    urlRequest.setValue("application/x-www-form-urlencoded", forHTTPHeaderField: "Content-Type")
-                    urlRequest.httpBody = formDataString.data(using: .utf8)
-                }
-
+            // Transform data in x-www-form-urlencoded
+            let formDataString = formData.map { key, value in
+                return "\(key)=\(value.addingPercentEncoding(withAllowedCharacters: .urlQueryAllowed) ?? "")"
+            }.joined(separator: "&")
+            urlRequest.setValue("application/x-www-form-urlencoded", forHTTPHeaderField: "Content-Type")
+            urlRequest.httpBody = formDataString.data(using: .utf8)
+        }
+        
         return urlRequest
     }
-
+    
     private func parse<T: Decodable>(data: Data, type _: T.Type, onResponse: @escaping (Result<T, Error>) -> Void) {
         do {
             let response = try decoder.decode(T.self, from: data)
