+++ FakeNFT/Helpers/UIColor+extension.swift
@@ -1,0 +1,19 @@
+//
+//  UIColor+extension.swift
+//  FakeNFT
+//
+//  Created by Nikita Tsomuk on 04.04.2024.
+//
+
+import UIKit
+
+extension UIColor {
+    static var nftBlack: UIColor { UIColor(named: "nftBlack") ?? UIColor.black }
+    static var nftGreen: UIColor { UIColor(named: "nftGreen") ?? UIColor.green }
+    static var nftRed: UIColor { UIColor(named: "nftRed") ?? UIColor.red }
+    static var nftBlue: UIColor { UIColor(named: "nftBlue") ?? UIColor.blue }
+    static var nftYellow: UIColor { UIColor(named: "nftYellow") ?? UIColor.yellow }
+    static var nftGray: UIColor { UIColor(named: "nftGray") ?? UIColor.darkGray }
+    static var nftLightGray: UIColor { UIColor(named: "nftLightGray") ?? UIColor.gray }
+    static var nftWhite: UIColor { UIColor(named: "nftWhite") ?? UIColor.white }
+}
