+++ FakeNFT/Helpers/UIImage+extension.swift
@@ -1,0 +1,20 @@
+//
+//  UIImage+extension.swift
+//  FakeNFT
+//
+//  Created by Nikita Tsomuk on 04.04.2024.
+//
+
+import UIKit
+
+extension UIImage {
+    //images
+    static var mockNft = UIImage(named: "mockNft") ?? UIImage()
+    //icons
+    static var basketDelete = UIImage(named: "basketDelete") ?? UIImage()
+    static var basketEmpty = UIImage(named: "basketEmpty") ?? UIImage()
+    static var basketFill = UIImage(named: "basketFill") ?? UIImage()
+    static var sort = UIImage(named: "sort") ?? UIImage()
+    static var done = UIImage(named: "Done") ?? UIImage()
+    static var close = UIImage(named: "close") ?? UIImage()
+}
